+++ components/MediaManager/RawHTML/index.js
@@ -1,0 +1,15 @@
+import { Wrapper, PointerEventControls } from "./styled";
+
+const RawHTML = ({ content }) => {
+  return (
+    <Wrapper tabIndex={0}>
+      <PointerEventControls
+        dangerouslySetInnerHTML={{
+          __html: content,
+        }}
+      />
+    </Wrapper>
+  );
+};
+
+export default RawHTML;