+++ three/components/PanoramaProjectionMesh/index.js
@@ -3,6 +3,7 @@ import * as THREE from "three";
 
 import { create3DRoom } from "../../core/RoomGeometry";
 import Shaders from "../../shaders";
+import { RENDER_ORDER } from "../../constant";
 
 const PanoramaProjectionMesh = ({
   three,
@@ -11,6 +12,7 @@ const PanoramaProjectionMesh = ({
   layout2D,
   panorama,
   panoramaOrigin,
+  onLoad,
 }) => {
   useEffect(() => {
     const { scene } = three;
@@ -20,6 +22,7 @@ const PanoramaProjectionMesh = ({
     const material = new THREE.ShaderMaterial({
       vertexShader: Shaders.vertexShaders.worldPosition,
       fragmentShader: Shaders.fragmentShaders.equirectangularProjection,
+      transparent: true,
     });
     Shaders.setUniforms.equirectangularProjection(material, {
       texture: panorama,
@@ -27,11 +30,15 @@ const PanoramaProjectionMesh = ({
     });
 
     const room = new THREE.Mesh(geometry, material);
+    room.renderOrder = RENDER_ORDER.MESH;
     scene.add(room);
+
+    onLoad(room);
+
     return () => {
       scene.remove(room);
     };
-  }, [three, floorY, ceilingY, layout2D, panorama, panoramaOrigin]);
+  }, [three, floorY, ceilingY, layout2D, panorama, panoramaOrigin, onLoad]);
 
   return null;
 };